Move conftest file, which was inexplicably located in the source folders ???
