+++ tests/data/diffusion/test_data_loader.py
@@ -4,9 +4,9 @@ from typing import Dict, List
 import pytest
 import torch
 
-from crystal_diffusion.data.diffusion.conftest import TestDiffusionDataBase
 from crystal_diffusion.data.diffusion.data_loader import (
     LammpsForDiffusionDataModule, LammpsLoaderParameters)
+from tests.conftest import TestDiffusionDataBase
 from tests.fake_data_utils import Configuration
